requirements.txt
# Core
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
pydantic==2.7.1
pyyaml==6.0.1

# PDF
pypdf==4.2.0

# Embeddings
sentence-transformers==3.0.1

# Vector DB
chromadb==0.5.0

# BM25
rank-bm25==0.2.2

# Reranking


# LLM clients
groq==0.9.0
anthropic==0.28.0

# Observability
langfuse==2.36.2
langsmith==0.1.77

# Evaluation
deepeval==1.4.3
ragas==0.1.14

# Experiment tracking
mlflow==2.13.0

# Dashboard
streamlit==1.35.0
plotly==5.22.0
pandas==2.2.2

# Utilities
httpx==0.27.0
tenacity>=8.3.0
loguru==0.7.2
